GP2-2437: fix test failures
This changeset remedies a failure in both tests from company/tests/test_templates.py

The bug appeared when directory-components was updated to support the V2/Magna header,
but the bug was not caused by the Magna work. Using git-bisect, it turned out to be
that the bug came from commit d3f7c8d1c93eee5a4b292c140e5fdf08681a1320
(https://github.com/uktrade/directory-components/commit/d3f7c8d1c93eee5a4b292c140e5fdf08681a1320)
which was for d-components v35.5.0

diretory-ui-buyer had previously been on directory-components==35.4.0, on which tests pass 100%

The bug itself was only present in tests, and came down to:
* directory-components 35.5.0 moved a template fragment into its own include, for code reuse
* that include made reference to a context variable (services_urls) which was not being provided
in the context passed to render_to_string() in the test
* the rendered content was failing to include the expected content because the entire include
was blowing up due to the absence of the context vars

Explicitly including the output of the `urls_processor` context processor fixed the bug.

Note that the rendered site was not affected by this change - only the tests